Add NullHandler to the root logger. Fixes #57
This was caused by trying to retrieve the description attribute of a pafy
object.
The description doesn't get loaded on object creation, so pafy makes a
HTTP call when the attribute gets accessed. Before that, if there is no
handler registered for the root logger, it adds a
streamhandler to it and sets the log level of the root logger
to debug.